# api-contract.yaml
openapi: 3.0.0
info:
  title: TarkVtark Debate Arena API
  description: |
    Backend API for TarkVtark debate application.
    This API manages debate topics, questions, replies, voting, and admin features.
    
    All data is stored in PostgreSQL database.
  version: 1.0.0
  contact:
    name: TarkVtark Team
    email: admin@tarkvtark.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.tarkvtark.com/api/v1
    description: Production server
...
paths:
...
  /admin/login:
    post:
      tags: [Admin]
      summary: Admin login
      description: Authenticate admin user and create session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/verify:
    post:
      tags: [Admin]
      summary: Verify JWT token
      description: Verify if the provided JWT token is valid and not expired
      parameters:
        - in: header
          name: Authorization
          required: true
          schema:
            type: string
          description: Bearer token
          example: "Bearer eyJhbGciOiJIUzUxMiJ9..."
      responses:
        '200':
          description: Token verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:
                    type: boolean
                    description: Whether the token is valid
                  username:
                    type: string
                    description: Username extracted from token (only if valid)
                  message:
                    type: string
                    description: Error message if token is invalid
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/guidelines:
    get:
      tags: [Admin]
      summary: Get guidelines
      description: Returns the list of community guidelines
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                example:
                  - "Be respectful and constructive in your arguments."
                  - "No hate speech, personal attacks, or discrimination."
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/faq:
    get:
      tags: [Admin]
      summary: Get FAQ items
      description: Returns the list of frequently asked questions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FAQItem'
        '500':
          $ref: '#/components/responses/ServerError'

  # =====================================================================
  # CONTACT ENDPOINTS
  # =====================================================================
...
components:
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 6
          maxLength: 100
    LoginResponse:
      type: object
      required:
        - success
        - user
      properties:
        success:
          type: boolean
        token:
          type: string
          nullable: true
          description: JWT token (if using token-based auth)
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
            username:
              type: string
            email:
              type: string
            fullName:
              type: string
    FAQItem:
      type: object
      required:
        - q
        - a
      properties:
        q:
          type: string
          description: The question
        a:
          type: string
          description: The answer

    # -----------------------------------------------------------------
    # CONTACT
    # -----------------------------------------------------------------
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          nullable: true
          description: Additional error details

  # =====================================================================
  # REUSABLE RESPONSES
  # =====================================================================
  responses:
...
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "UNAUTHORIZED"
            message: "Invalid credentials"
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "INTERNAL_ERROR"
            message: "An unexpected error occurred"